refactor: разнести логику класса EntityExistChecker на разные адаптеры, исправить тесты под внесенные изменения
