+++ src/main/java/faang/school/postservice/util/post/PostRepositoryAdapter.java
@@ -1,0 +1,23 @@
+package faang.school.postservice.util.post;
+
+import faang.school.postservice.exception.EntityNotFoundException;
+import faang.school.postservice.model.Post;
+import faang.school.postservice.repository.PostRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+@Slf4j
+public class PostRepositoryAdapter {
+
+    private final PostRepository postRepository;
+
+    public Post getPostById(long postId) {
+        return postRepository.findById(postId).orElseThrow(() -> {
+            log.error("Post #{} doesn't exist", postId);
+            return new EntityNotFoundException("Post doesn't exist");
+        });
+    }
+}
